requirements.txt
openai
tiktoken
pandas
numpy
pyarrow
matplotlib
seaborn
plotly
termcolor
newspaper3k
pinecone-client
google-generativeai
google-genai
groq
ollama
anthropic
mistralai
flask
ipython
nbformat
python-dotenv
lxml[html_clean]
gunicorn
statsmodels
scipy
scikit-learn
contextily
geopandas
yfinance